second try

+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: tkoami <tkoami@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/02 15:53:25 by tkoami            #+#    #+#             */
-/*   Updated: 2020/11/20 15:28:44 by tkoami           ###   ########.fr       */
+/*   Updated: 2020/11/25 21:14:29 by tkoami           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,11 +15,11 @@
 void	*ft_memset(void *b, int c, size_t len)
 {
 	unsigned char	*tmp;
-	int				i;
+	size_t			i;
 
 	tmp = (unsigned char*)b;
 	i = 0;
-	while (i < (int)len)
+	while (i < len)
 	{
 		tmp[i] = (unsigned char)c;
 		i++;
